requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
python-docx
openai
tiktoken
python-dotenv
